requirements.txt
streamlit
torch
transformers
datasets
gdown
scikit-learn
pandas
python-dotenv
google-genai